Use source countries table for all suggests
Currently, when trying to suggest on e.g. `countries.names`, Looker tries to query the entire source dataset for distinct field names. The problem is that the source datasets (e.g. `active_users_aggregates`) usually require a `submission_date` filter. Because the suggest doesn't filter, the suggest fails, and nothing shows.

This is visible in e.g. [this dashboard](https://mozilla.cloud.looker.com/dashboards/885?Country=).

This change will query _just_ the `countries` table for distinct country names and codes. This is a bit of a behavior change, but since suggests by and large were not working previously, I believe these are acceptable: Countries will present themselves in these results even if the underlying dataset doesn't have that country present.