pause KPI reporting while we're updating active_users_aggregates
We are pushing changes to active_users_aggregates that will update the DAU definitions for Desktop and Mobile. This will cause qualitative changes to the historical and forecasted data.

While forecast differences due to the new definition changes have already been presented to Firefox Leadership, the changes may be surprising to dashboard viewers who do not already understand the nuances of the definitions. This freeze will enable us to communicate that changes are coming to the dashboard to a broader audience.